finished project

+++ Project0/src/main/java/com/revature/util/userInput/InputValidation.java
@@ -2,11 +2,11 @@ package com.revature.util.userInput;
 
 public class InputValidation {
 
-	public static boolean ValidateValue(int value){
+	public static boolean ValidateValue(double value){
 		return value > 0;
 	}	
 	
-	public static boolean OverdrawCheck(int value) {
+	public static boolean OverdrawCheck(double value) {
 		return GetInput.currentBankAccount.getAccountBalance()- value > 0;
 		
 	}
